# docs/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Media Collection API",
    "description": "A simple API for managing a media collection (movies, TV shows, books, etc.)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Server"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication endpoints"
    },
    {
      "name": "Users",
      "description": "User profile endpoints"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Better-Auth session cookie"
      }
    }
  },
  "paths": {
    "/openapi": {
      "get": {
        "operationId": "getOpenapi",
        "responses": {
          "200": {}
        }
      }
    },
    "/docs": {
      "get": {
        "operationId": "getDocs",
        "responses": {
          "200": {}
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {}
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "postApiAuthRegister",
        "tags": [
          "Auth"
        ],
        "description": "Register a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "description": "User email address",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "description": "User password (minimum 8 characters)"
                  },
                  "name": {
                    "description": "User display name",
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {},
                    "session": {},
                    "sessionToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "example": {
                  "message": "Registration successful"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or content type"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "postApiAuthLogin",
        "tags": [
          "Auth"
        ],
        "description": "Log in with credentials and receive a session cookie",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "description": "User email address",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "description": "User password"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {},
                    "session": {},
                    "sessionToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "example": {
                  "message": "Login successful"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "Session cookie set by Better-Auth",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid content type"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "postApiAuthLogout",
        "tags": [
          "Auth"
        ],
        "description": "Log out and revoke the session cookie",
        "parameters": [
          {
            "name": "better-auth.session_token",
            "in": "cookie",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Better-Auth session cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Logout result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {},
                    "session": {},
                    "sessionToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "example": {
                  "message": "Logout successful"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "postApiAuthForgotPassword",
        "tags": [
          "Auth"
        ],
        "description": "Initiate password reset process by sending a reset email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "description": "User email address",
                    "example": "user@example.com"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {},
                    "session": {},
                    "sessionToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "example": {
                  "message": "Password reset email sent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or content type"
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "postApiAuthResetPassword",
        "tags": [
          "Auth"
        ],
        "description": "Complete password reset process by providing new password and reset token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Password reset token"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 8,
                    "description": "New password (minimum 8 characters)"
                  }
                },
                "required": [
                  "token",
                  "newPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {},
                    "session": {},
                    "sessionToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "example": {
                  "message": "Password reset successful"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or content type"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "getApiAuthMe",
        "tags": [
          "Auth"
        ],
        "description": "Get the authenticated user profile and session info",
        "parameters": [
          {
            "name": "better-auth.session_token",
            "in": "cookie",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Better-Auth session cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user profile and session info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {},
                    "session": {},
                    "sessionToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "example": {
                  "message": "Authenticated user profile and session info",
                  "user": {
                    "id": "user_123",
                    "email": "user@example.com"
                  },
                  "session": {
                    "id": "session_123",
                    "expiresAt": "2026-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/profile": {
      "get": {
        "operationId": "getApiUsersProfile",
        "tags": [
          "Users"
        ],
        "description": "Get the authenticated user profile",
        "parameters": [
          {
            "name": "better-auth.session",
            "in": "cookie",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Better-Auth session cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {}
                  },
                  "required": [
                    "user"
                  ]
                },
                "example": {
                  "user": {
                    "id": "user_123",
                    "email": "user@example.com",
                    "name": "User Example",
                    "username": "userexample"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "patch": {
        "operationId": "patchApiUsersProfile",
        "tags": [
          "Users"
        ],
        "description": "Update authenticated user profile",
        "parameters": [
          {
            "name": "better-auth.session",
            "in": "cookie",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Better-Auth session cookie"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "image": {
                    "type": "string",
                    "minLength": 1
                  },
                  "username": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 40
                  },
                  "displayUsername": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {}
                  },
                  "required": [
                    "user"
                  ]
                },
                "example": {
                  "user": {
                    "id": "user_123",
                    "email": "user@example.com",
                    "name": "Updated Name",
                    "username": "newusername",
                    "displayUsername": "New Username"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "getApiUsersById",
        "tags": [
          "Users"
        ],
        "description": "Get a public user profile by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {}
                  },
                  "required": [
                    "user"
                  ]
                },
                "example": {
                  "user": {
                    "id": "user_123",
                    "email": "user@example.com",
                    "name": "User Example"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  }
}